[#1] refactor: ProducItem 분리
